+++ packages/go_router_builder/example/lib/all_types.dart
@@ -30,7 +30,8 @@ class AllTypesBaseRoute extends GoRouteData {
   const AllTypesBaseRoute();
 
   @override
-  Widget build(BuildContext context) => const BasePage<void>(
+  Widget build(BuildContext context, GoRouterState state) =>
+      const BasePage<void>(
         dataTitle: 'Root',
         param: null,
       );
@@ -46,7 +47,7 @@ class BigIntRoute extends GoRouteData {
   final BigInt? bigIntField;
 
   @override
-  Widget build(BuildContext context) => BasePage<BigInt>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<BigInt>(
         dataTitle: 'BigIntRoute',
         param: requiredBigIntField,
         queryParam: bigIntField,
@@ -71,7 +72,7 @@ class BoolRoute extends GoRouteData {
   final bool boolFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<bool>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<bool>(
         dataTitle: 'BoolRoute',
         param: requiredBoolField,
         queryParam: boolField,
@@ -95,7 +96,7 @@ class DateTimeRoute extends GoRouteData {
   final DateTime? dateTimeField;
 
   @override
-  Widget build(BuildContext context) => BasePage<DateTime>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<DateTime>(
         dataTitle: 'DateTimeRoute',
         param: requiredDateTimeField,
         queryParam: dateTimeField,
@@ -120,7 +121,7 @@ class DoubleRoute extends GoRouteData {
   final double doubleFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<double>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<double>(
         dataTitle: 'DoubleRoute',
         param: requiredDoubleField,
         queryParam: doubleField,
@@ -146,7 +147,7 @@ class IntRoute extends GoRouteData {
   final int intFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<int>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<int>(
         dataTitle: 'IntRoute',
         param: requiredIntField,
         queryParam: intField,
@@ -172,7 +173,7 @@ class NumRoute extends GoRouteData {
   final num numFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<num>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<num>(
         dataTitle: 'NumRoute',
         param: requiredNumField,
         queryParam: numField,
@@ -198,7 +199,8 @@ class EnumRoute extends GoRouteData {
   final PersonDetails enumFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<PersonDetails>(
+  Widget build(BuildContext context, GoRouterState state) =>
+      BasePage<PersonDetails>(
         dataTitle: 'EnumRoute',
         param: requiredEnumField,
         queryParam: enumField,
@@ -224,7 +226,8 @@ class EnhancedEnumRoute extends GoRouteData {
   final SportDetails enumFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<SportDetails>(
+  Widget build(BuildContext context, GoRouterState state) =>
+      BasePage<SportDetails>(
         dataTitle: 'EnhancedEnumRoute',
         param: requiredEnumField,
         queryParam: enumField,
@@ -250,7 +253,7 @@ class StringRoute extends GoRouteData {
   final String stringFieldWithDefaultValue;
 
   @override
-  Widget build(BuildContext context) => BasePage<String>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<String>(
         dataTitle: 'StringRoute',
         param: requiredStringField,
         queryParam: stringField,
@@ -274,7 +277,7 @@ class UriRoute extends GoRouteData {
   final Uri? uriField;
 
   @override
-  Widget build(BuildContext context) => BasePage<Uri>(
+  Widget build(BuildContext context, GoRouterState state) => BasePage<Uri>(
         dataTitle: 'UriRoute',
         param: requiredUriField,
         queryParam: uriField,