+++ packages/pigeon/lib/kotlin_generator.dart
@@ -4,6 +4,7 @@
 
 import 'ast.dart';
 import 'functional.dart';
+import 'generator.dart';
 import 'generator_tools.dart';
 import 'pigeon_lib.dart' show TaskQueueType;
 
@@ -64,6 +65,20 @@ class KotlinOptions {
   }
 }
 
+/// Class that manages all Kotlin code generation.
+class KotlinGenerator extends Generator<KotlinOptions> {
+  /// Instantiates a Kotlin Generator.
+  KotlinGenerator();
+
+  /// Generates Kotlin files with specified [KotlinOptions]
+  @override
+  void generate(KotlinOptions languageOptions, Root root, StringSink sink,
+      {FileType fileType = FileType.na}) {
+    assert(fileType == FileType.na);
+    generateKotlin(languageOptions, root, sink);
+  }
+}
+
 /// Calculates the name of the codec that will be generated for [api].
 String _getCodecName(Api api) => '${api.name}Codec';
 
@@ -85,10 +100,9 @@ void _writeCodec(Indent indent, Api api, Root root) {
         for (final EnumeratedClass customClass in codecClasses) {
           indent.write('${customClass.enumeration}.toByte() -> ');
           indent.scoped('{', '}', () {
-            indent.write(
-                'return (readValue(buffer) as? Map<String, Any?>)?.let ');
+            indent.write('return (readValue(buffer) as? List<Any?>)?.let ');
             indent.scoped('{', '}', () {
-              indent.writeln('${customClass.name}.fromMap(it)');
+              indent.writeln('${customClass.name}.fromList(it)');
             });
           });
         }
@@ -105,7 +119,7 @@ void _writeCodec(Indent indent, Api api, Root root) {
           indent.write('is ${customClass.name} -> ');
           indent.scoped('{', '}', () {
             indent.writeln('stream.write(${customClass.enumeration})');
-            indent.writeln('writeValue(stream, value.toMap())');
+            indent.writeln('writeValue(stream, value.toList())');
           });
         }
         indent.writeln('else -> super.writeValue(stream, value)');
@@ -214,8 +228,7 @@ void _writeHostApi(Indent indent, Api api, Root root) {
 
               indent.write('channel.setMessageHandler ');
               indent.scoped('{ $messageVarName, reply ->', '}', () {
-                indent.writeln('val wrapped = hashMapOf<String, Any?>()');
-
+                indent.writeln('var wrapped = listOf<Any?>()');
                 indent.write('try ');
                 indent.scoped('{', '}', () {
                   final List<String> methodArgument = <String>[];
@@ -233,26 +246,21 @@ void _writeHostApi(Indent indent, Api api, Root root) {
                       'api.${method.name}(${methodArgument.join(', ')})';
                   if (method.isAsynchronous) {
                     indent.write('$call ');
-                    if (method.returnType.isVoid) {
-                      indent.scoped('{', '}', () {
-                        indent.writeln('reply.reply(null)');
-                      });
-                    } else {
-                      indent.scoped('{', '}', () {
-                        indent.writeln('reply.reply(wrapResult(it))');
-                      });
-                    }
+                    final String resultValue =
+                        method.returnType.isVoid ? 'null' : 'it';
+                    indent.scoped('{', '}', () {
+                      indent.writeln('reply.reply(wrapResult($resultValue))');
+                    });
                   } else if (method.returnType.isVoid) {
                     indent.writeln(call);
-                    indent.writeln('wrapped["${Keys.result}"] = null');
+                    indent.writeln('wrapped = listOf<Any?>(null)');
                   } else {
-                    indent.writeln('wrapped["${Keys.result}"] = $call');
+                    indent.writeln('wrapped = listOf<Any?>($call)');
                   }
                 }, addTrailingNewline: false);
                 indent.add(' catch (exception: Error) ');
                 indent.scoped('{', '}', () {
-                  indent.writeln(
-                      'wrapped["${Keys.error}"] = wrapError(exception)');
+                  indent.writeln('wrapped = wrapError(exception)');
                   if (method.isAsynchronous) {
                     indent.writeln('reply.reply(wrapped)');
                   }
@@ -478,20 +486,16 @@ void generateKotlin(KotlinOptions options, Root root, StringSink sink) {
         indent, anEnum.documentationComments, _docCommentSpec);
     indent.write('enum class ${anEnum.name}(val raw: Int) ');
     indent.scoped('{', '}', () {
-      // We use explicit indexing here as use of the ordinal() method is
-      // discouraged. The toMap and fromMap API matches class API to allow
-      // the same code to work with enums and classes, but this
-      // can also be done directly in the host and flutter APIs.
-      int index = 0;
-      for (final String member in anEnum.members) {
-        indent.write('${member.toUpperCase()}($index)');
+      enumerate(anEnum.members, (int index, final EnumMember member) {
+        addDocumentationComments(
+            indent, member.documentationComments, _docCommentSpec);
+        indent.write('${member.name.toUpperCase()}($index)');
         if (index != anEnum.members.length - 1) {
           indent.addln(',');
         } else {
           indent.addln(';');
         }
-        index++;
-      }
+      });
 
       indent.writeln('');
       indent.write('companion object ');
@@ -514,48 +518,41 @@ void generateKotlin(KotlinOptions options, Root root, StringSink sink) {
       indent.add(defaultNil);
     }
 
-    void writeToMap() {
-      indent.write('fun toMap(): Map<String, Any?> ');
+    void writeToList() {
+      indent.write('fun toList(): List<Any?> ');
       indent.scoped('{', '}', () {
-        indent.writeln('val map = mutableMapOf<String, Any?>()');
-
-        for (final NamedType field in klass.fields) {
-          final HostDatatype hostDatatype = getHostDatatype(field);
-          String toWriteValue = '';
-          final String fieldName = field.name;
-          final String prefix = field.type.isNullable ? 'it' : fieldName;
-          if (!hostDatatype.isBuiltin &&
-              rootClassNameSet.contains(field.type.baseName)) {
-            toWriteValue = '$prefix.toMap()';
-          } else if (!hostDatatype.isBuiltin &&
-              rootEnumNameSet.contains(field.type.baseName)) {
-            toWriteValue = '$prefix.raw';
-          } else {
-            toWriteValue = prefix;
+        indent.write('return listOf<Any?>');
+        indent.scoped('(', ')', () {
+          for (final NamedType field in getFieldsInSerializationOrder(klass)) {
+            final HostDatatype hostDatatype = getHostDatatype(field);
+            String toWriteValue = '';
+            final String fieldName = field.name;
+            if (!hostDatatype.isBuiltin &&
+                rootClassNameSet.contains(field.type.baseName)) {
+              toWriteValue = '$fieldName?.toList()';
+            } else if (!hostDatatype.isBuiltin &&
+                rootEnumNameSet.contains(field.type.baseName)) {
+              toWriteValue = '$fieldName?.raw';
+            } else {
+              toWriteValue = fieldName;
+            }
+            indent.writeln('$toWriteValue,');
           }
-
-          if (field.type.isNullable) {
-            indent.writeln(
-                '$fieldName?.let { map["${field.name}"] = $toWriteValue }');
-          } else {
-            indent.writeln('map["${field.name}"] = $toWriteValue');
-          }
-        }
-
-        indent.writeln('return map');
+        });
       });
     }
 
-    void writeFromMap() {
+    void writeFromList() {
       final String className = klass.name;
 
       indent.write('companion object ');
       indent.scoped('{', '}', () {
         indent.writeln('@Suppress("UNCHECKED_CAST")');
-        indent.write('fun fromMap(map: Map<String, Any?>): $className ');
+        indent.write('fun fromList(list: List<Any?>): $className ');
 
         indent.scoped('{', '}', () {
-          for (final NamedType field in klass.fields) {
+          enumerate(getFieldsInSerializationOrder(klass),
+              (int index, final NamedType field) {
             final HostDatatype hostDatatype = getHostDatatype(field);
 
             // The StandardMessageCodec can give us [Integer, Long] for
@@ -563,50 +560,51 @@ void generateKotlin(KotlinOptions options, Root root, StringSink sink) {
             // longs in Pigeon with Kotlin.
             final bool isInt = field.type.baseName == 'int';
 
-            final String mapValue = 'map["${field.name}"]';
+            final String listValue = 'list[$index]';
             final String fieldType = _kotlinTypeForDartType(field.type);
 
             if (field.type.isNullable) {
               if (!hostDatatype.isBuiltin &&
                   rootClassNameSet.contains(field.type.baseName)) {
                 indent.write('val ${field.name}: $fieldType? = ');
-                indent.add('($mapValue as? Map<String, Any?>)?.let ');
+                indent.add('($listValue as? List<Any?>)?.let ');
                 indent.scoped('{', '}', () {
-                  indent.writeln('$fieldType.fromMap(it)');
+                  indent.writeln('$fieldType.fromList(it)');
                 });
               } else if (!hostDatatype.isBuiltin &&
                   rootEnumNameSet.contains(field.type.baseName)) {
                 indent.write('val ${field.name}: $fieldType? = ');
-                indent.add('($mapValue as? Int)?.let ');
+                indent.add('($listValue as? Int)?.let ');
                 indent.scoped('{', '}', () {
                   indent.writeln('$fieldType.ofRaw(it)');
                 });
               } else if (isInt) {
-                indent.write('val ${field.name} = $mapValue');
+                indent.write('val ${field.name} = $listValue');
                 indent.addln(
                     '.let { if (it is Int) it.toLong() else it as? Long }');
               } else {
-                indent.writeln('val ${field.name} = $mapValue as? $fieldType');
+                indent.writeln('val ${field.name} = $listValue as? $fieldType');
               }
             } else {
               if (!hostDatatype.isBuiltin &&
                   rootClassNameSet.contains(field.type.baseName)) {
                 indent.writeln(
-                    'val ${field.name} = $fieldType.fromMap($mapValue as Map<String, Any?>)');
+                    'val ${field.name} = $fieldType.fromList($listValue as List<Any?>)');
               } else if (!hostDatatype.isBuiltin &&
                   rootEnumNameSet.contains(field.type.baseName)) {
                 indent.write(
-                    'val ${field.name} = $fieldType.ofRaw($mapValue as Int)!!');
+                    'val ${field.name} = $fieldType.ofRaw($listValue as Int)!!');
               } else {
-                indent.writeln('val ${field.name} = $mapValue as $fieldType');
+                indent.writeln('val ${field.name} = $listValue as $fieldType');
               }
             }
-          }
+          });
 
           indent.writeln('');
           indent.write('return $className(');
-          for (final NamedType field in klass.fields) {
-            final String comma = klass.fields.last == field ? '' : ', ';
+          for (final NamedType field in getFieldsInSerializationOrder(klass)) {
+            final String comma =
+                getFieldsInSerializationOrder(klass).last == field ? '' : ', ';
             indent.add('${field.name}$comma');
           }
           indent.addln(')');
@@ -623,9 +621,9 @@ void generateKotlin(KotlinOptions options, Root root, StringSink sink) {
 
     indent.write('data class ${klass.name} ');
     indent.scoped('(', '', () {
-      for (final NamedType element in klass.fields) {
+      for (final NamedType element in getFieldsInSerializationOrder(klass)) {
         writeField(element);
-        if (klass.fields.last != element) {
+        if (getFieldsInSerializationOrder(klass).last != element) {
           indent.addln(',');
         } else {
           indent.addln('');
@@ -634,8 +632,8 @@ void generateKotlin(KotlinOptions options, Root root, StringSink sink) {
     });
 
     indent.scoped(') {', '}', () {
-      writeFromMap();
-      writeToMap();
+      writeFromList();
+      writeToList();
     });
   }
 
@@ -648,26 +646,21 @@ void generateKotlin(KotlinOptions options, Root root, StringSink sink) {
   }
 
   void writeWrapResult() {
-    indent.write('private fun wrapResult(result: Any?): Map<String, Any?> ');
+    indent.write('private fun wrapResult(result: Any?): List<Any?> ');
     indent.scoped('{', '}', () {
-      indent.writeln('return hashMapOf("result" to result)');
+      indent.writeln('return listOf(result)');
     });
   }
 
   void writeWrapError() {
-    indent.write(
-        'private fun wrapError(exception: Throwable): Map<String, Any> ');
+    indent.write('private fun wrapError(exception: Throwable): List<Any> ');
     indent.scoped('{', '}', () {
       indent.write('return ');
-      indent.scoped('hashMapOf<String, Any>(', ')', () {
-        indent.write('"error" to ');
-        indent.scoped('hashMapOf<String, Any>(', ')', () {
-          indent.writeln(
-              '"${Keys.errorCode}" to exception.javaClass.simpleName,');
-          indent.writeln('"${Keys.errorMessage}" to exception.toString(),');
-          indent.writeln(
-              '"${Keys.errorDetails}" to "Cause: " + exception.cause + ", Stacktrace: " + Log.getStackTraceString(exception)');
-        });
+      indent.scoped('listOf<Any>(', ')', () {
+        indent.writeln('exception.javaClass.simpleName,');
+        indent.writeln('exception.toString(),');
+        indent.writeln(
+            '"Cause: " + exception.cause + ", Stacktrace: " + Log.getStackTraceString(exception)');
       });
     });
   }