+++ packages/pigeon/platform_tests/test_plugin/windows/test/multiple_arity_test.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include "test/multiple_arity.g.h"
+#include "pigeon/multiple_arity.gen.h"
 #include "test/utils/fake_host_messenger.h"
 
 namespace multiple_arity_pigeontest {
@@ -25,7 +25,7 @@ class TestHostApi : public MultipleArityHostApi {
 };
 
 const EncodableValue& GetResult(const EncodableValue& pigeon_response) {
-  return std::get<EncodableMap>(pigeon_response).at(EncodableValue("result"));
+  return std::get<EncodableList>(pigeon_response)[0];
 }
 }  // namespace